+++ elements/itemfilter/src/filters/text.ts
@@ -16,6 +16,7 @@ export class EOxItemFilterText extends LitElement {
       this.renderRoot.querySelector("input[type='text']")
     );
     this.filterObject.keys.forEach((key) => {
+      // @ts-ignore
       this.filterObject.state[key] = searchInput.value;
     });
     this.filterObject.dirty = true;