+++ elements/itemfilter/src/global.d.ts
@@ -1,12 +1,12 @@
 declare type Item = {
-  [key: string]: any;
+  [key: string]: unknown;
 };
 
 declare type Filter = HTMLElement & {
-  reset: Function;
+  reset: () => void;
 };
 
-declare type FilterState = { [key: string]: any };
+declare type FilterState = { [key: string]: object };
 
 declare type FilterObject = {
   dirty?: boolean;
@@ -14,7 +14,8 @@ declare type FilterObject = {
   expanded?: boolean;
   featured?: boolean;
   format?: string;
-  key: string;
+  key?: string;
+  keys?: string[];
   reset: Function;
   state?: FilterState;
   stringifiedState?: string;