+++ elements/itemfilter/test/state.cy.ts
@@ -10,21 +10,23 @@ describe("Item Filter Config", () => {
   beforeEach(() => {
     cy.mount(`<eox-itemfilter></eox-itemfilter>`)
       .as("eox-itemfilter")
-      .then((eoxItemFilter: any) => {
-        eoxItemFilter[0].config = {
+      .then(($el) => {
+        const eoxItemFilter = <EOxItemFilter>$el[0];
+        eoxItemFilter.config = {
           titleProperty: "title",
           filterProperties: [
             {
               key: "themes",
               type: "multiselect",
               expanded: true,
+              // @ts-ignore
               state,
             },
           ],
           aggregateResults: "themes",
         };
-        eoxItemFilter[0].apply(testItems);
-        eoxItemFilter[0].selectedResult = testItems[selectedResultIndex];
+        eoxItemFilter.apply(testItems);
+        eoxItemFilter.selectedResult = testItems[selectedResultIndex];
       });
   });