+++ elements/layercontrol/src/components/layerTools.js
@@ -1,4 +1,4 @@
-import { LitElement, html } from "lit";
+import { LitElement, html, nothing } from "lit";
 import { unsafeHTML } from "lit/directives/unsafe-html.js";
 import { map } from "lit/directives/map.js";
 import { when } from "lit/directives/when.js";
@@ -20,6 +20,7 @@ export class EOxLayerControlLayerTools extends LitElement {
     layer: { attribute: false },
     tools: { attribute: false },
     unstyled: { type: Boolean },
+    noShadow: { type: Boolean },
   };
 
   constructor() {
@@ -41,6 +42,11 @@ export class EOxLayerControlLayerTools extends LitElement {
      * Render the element without additional styles
      */
     this.unstyled = false;
+
+    /**
+     * Renders the element without a shadow root
+     */
+    this.noShadow = true;
   }
 
   /**
@@ -90,16 +96,18 @@ export class EOxLayerControlLayerTools extends LitElement {
         );
       }}
     >
-      x
+      ${this.unstyled ? "x" : nothing}
     </button>
   `;
 
   _sortButton = html`
-    <button class="sort-icon icon drag-handle">sort</button>
+    <button class="sort-icon icon drag-handle">
+      ${this.unstyled ? "sort" : nothing}
+    </button>
   `;
 
   createRenderRoot() {
-    return this;
+    return this.noShadow ? this : super.createRenderRoot();
   }
 
   render() {
@@ -127,7 +135,10 @@ export class EOxLayerControlLayerTools extends LitElement {
               </div>
             `,
             () => html`
-              <details class="tools">
+              <details
+                class="tools"
+                open=${this.layer.get("layerControlToolsExpand") || nothing}
+              >
                 <summary>
                   <button
                     class="icon ${this.tools.length === 1
@@ -138,6 +149,7 @@ export class EOxLayerControlLayerTools extends LitElement {
                   </button>
                 </summary>
                 <eox-layercontrol-tabs
+                  .noShadow=${false}
                   .actions=${this._parseActions(this.tools)}
                   .tabs=${this._parseTools(this.tools)}
                   .unstyled=${this.unstyled}
@@ -145,7 +157,9 @@ export class EOxLayerControlLayerTools extends LitElement {
                   ${map(
                     this._parseTools(this.tools),
                     (tool) => html`
-                      <button slot="${tool}-icon" class="icon">${tool}</button>
+                      <button slot="${tool}-icon" class="icon">
+                        ${this.unstyled ? tool : nothing}
+                      </button>
                     `
                   )}