+++ elements/map/test/utils/events.ts
@@ -14,14 +14,16 @@ export function simulateEvent(
   const viewport = map.getViewport();
   // calculated in case body has top < 0 (test runner with small window)
   const position = viewport.getBoundingClientRect();
-  const event = {} as any;
-  event.type = type;
-  event.target = viewport.firstChild;
-  event.clientX = position.left + x + width / 2;
-  event.clientY = position.top + y + height / 2;
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  event.preventDefault = function () {};
-  event.pointerType = "mouse";
+  const event = {
+    type: type,
+    target: viewport.firstChild,
+    clientX: position.left + x + width / 2,
+    clientY: position.top + y + height / 2,
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    preventDefault: function () {},
+    pointerType: "mouse",
+  };
+  // @ts-ignore
   const simulatedEvent = new MapBrowserEvent(type, map, event);
   map.handleMapBrowserEvent(simulatedEvent);
   return simulatedEvent;